YU-RIS: Add support for yscfg.ybn, hook GetTempFileNameA()
